move NN_NAMESPACE map to connect
