+++ src/stream_ml/core/_multi/bases.py
@@ -10,7 +10,7 @@ from dataclasses import KW_ONLY, dataclass, fields
 from typing import TYPE_CHECKING, Any, Protocol
 
 from stream_ml.core._core.api import Model
-from stream_ml.core._core.base import NN_NAMESPACE
+from stream_ml.core._connect.nn_namespace import NN_NAMESPACE
 from stream_ml.core.setup_package import CompiledShim
 from stream_ml.core.typing import (
     Array,
